Fixed an issue with passing an interpolation mapping in translator.
Previously it was possible to provide an interpolation mapping only as part of
a translation string. Providing it later during translation would be ignored.

This change allows you to provide the mapping as part of the translator call,
e.g.

  translator(tstring, mapping={...})

if the tstring does not contain a mapping itself, the mapping given in the
translator() call will be used as is. Otherwise, the previously given mapping
will be updated with the new values allowing partial overriding of an existing
mapping.